+++ packages/zephyr-react/src/components/Typography/index.ts
@@ -1,2 +1,3 @@
 export type { TypographyProps } from './Typography';
+export { H1, H2, P, Span } from './Typography';
 export { default } from './Typography';