+++ packages/zephyr-react/src/types/index.ts
@@ -1,6 +1,8 @@
 import type { Argument } from 'classnames';
+import type { ClassValue } from 'clsx';
 import type { ReactNode } from 'react';
 
+export type Customization = ClassValue | ClassValue[] | [...ClassValue[], boolean];
 export type ClassName = string | { overrides: Argument[] };
 
 export type DisplayValueObject = {